Create weighted maze.py
Input
4
1,8,21,7
19,17,10,20
2,18,23,22
14,25,4,13
Output
22

Explanation
One possible path through the maze is through the intersections with coordinates
(1,1),(1,2),(2,2),(2,3)(2,4)(3,4),(4,4). This would result in the cart having a weight of 22 kg ( at the intersections on this path,, the weight in the cart after the exchange are 1,8,17,18,18,20,22,22

Input
5 1,29,40,24,12 13,31,40,31,33 29,40,17,35,32 15,39,28,3,31 15,21,31,38,24
Output
38
Explanation
One path through the maze is (1,1),(2,1), (3,1),(4,1),(5,1),(5,2), (5,3), (5,4) (5,5). The maximum weight in this path is 38, which is the weight in the cart when leaving the maze.
As there is no other pathe in the maze which has a lower weight in the cart when exiting it, the output is 38